+++ include/de_bruijn/node.h
@@ -8,12 +8,12 @@
 
 class debruijn::Node {
 public:
-    uint16_t id;
+    uint32_t id;
     std::deque<uint16_t> hashed_node_ids;
     std::unordered_multiset<uint32_t> read_ids;
-    std::unordered_set<uint16_t> out_nodes;
+    std::unordered_set<uint32_t> out_nodes;
 
-    Node(const uint16_t, const deque<uint16_t>&, const uint32_t);
+    Node(const uint32_t, const deque<uint16_t>&, const uint32_t);
 
     bool operator == (const Node& y) const;
     bool operator != (const Node& y) const;
