+++ include/noise_filtering.h
@@ -11,7 +11,7 @@ void hashed_node_ids_to_ids_and_orientations(const deque<uint16_t>&, std::vector
 bool overlap_forwards(const deque<uint16_t>&, const deque<uint16_t>&);
 bool overlap_backwards(const deque<uint16_t>&, const deque<uint16_t>&);
 deque<uint16_t> reverse_hashed_node(const deque<uint16_t>&);
-void dbg_node_ids_to_ids_and_orientations(const debruijn::Graph&, const deque<uint16_t>&, std::vector<uint16_t>&, std::vector<bool>&);
+void dbg_node_ids_to_ids_and_orientations(const debruijn::Graph&, const deque<uint32_t>&, std::vector<uint16_t>&, std::vector<bool>&);
 debruijn::Graph construct_debruijn_graph_from_pangraph(uint8_t, const pangenome::Graph*);
 void remove_leaves(pangenome::Graph*, debruijn::Graph &, uint16_t covg_thresh=1);
 void filter_unitigs(pangenome::Graph*, debruijn::Graph &, const uint16_t&);
